correccion de errores

+++ src/AgoraHub360.ERP.Application/Services/CategoryService.cs
@@ -25,7 +25,7 @@ public class CategoryService : ICategoryService
 
     public async Task<Result<CategoryDto>> GetByIdAsync(long id, CancellationToken ct = default)
     {
-        var entity = await _repo.GetByIdAsync(id, ct);
+        var entity = await _repo.GetByIdAsync((int)id, ct);
         if (entity is null) return Result<CategoryDto>.Failure($"Categoría {id} no encontrada.");
         return Result<CategoryDto>.Success(Map(entity));
     }
@@ -54,7 +54,7 @@ public class CategoryService : ICategoryService
 
     public async Task<Result<CategoryDto>> UpdateAsync(long id, UpdateCategoryDto dto, CancellationToken ct = default)
     {
-        var entity = await _repo.GetByIdAsync(id, ct);
+        var entity = await _repo.GetByIdAsync((int)id, ct);
         if (entity is null) return Result<CategoryDto>.Failure($"Categoría {id} no encontrada.");
 
         entity.ParentCategoryId = dto.ParentCategoryId;
@@ -69,7 +69,7 @@ public class CategoryService : ICategoryService
 
     public async Task<Result<bool>> DeleteAsync(long id, CancellationToken ct = default)
     {
-        var entity = await _repo.GetByIdAsync(id, ct);
+        var entity = await _repo.GetByIdAsync((int)id, ct);
         if (entity is null) return Result<bool>.Failure($"Categoría {id} no encontrada.");
         await _repo.DeleteAsync(entity, ct);
         await _uow.SaveChangesAsync(ct);
